Add SearchHeroForm, SearchBlockHero and SearchInputLabel components
