+++ app/components/atoms/SearchBlockHero/index.js
@@ -1,0 +1,8 @@
+import styled from 'styled-components';
+
+const SearchBlockHero = styled.div`
+  width: 100%;
+  background: ${props => props.theme.color.brand.primary};
+`;
+
+export default SearchBlockHero;
