+++ app/components/molecules/SearchHeroForm/index.js
@@ -1,0 +1,12 @@
+import styled from 'styled-components';
+
+const SearchHeroForm = styled.form`
+  position: relative;
+  max-width: 400px;
+  margin: auto;
+  padding-top: 1rem;
+  padding-bottom: 2rem;
+  display: flex;
+`;
+
+export default SearchHeroForm;
