+++ app/pages/search/content.js
@@ -15,6 +15,9 @@ import PageTitle from '../../components/atoms/PageTitle';
 import SearchIcon from '../../components/atoms/SearchIcon';
 import MenuItem from '../../components/atoms/MenuItem';
 import SecondaryNavigation from '../../components/molecules/SecondaryNavigation';
+import SearchInputLabel from '../../components/atoms/SearchInputLabel';
+import SearchBlockHero from '../../components/atoms/SearchBlockHero';
+import SearchHeroForm from '../../components/molecules/SearchHeroForm';
 
 const SearchContainer = styled.div`
   margin: auto;
@@ -22,20 +25,6 @@ const SearchContainer = styled.div`
   max-width: ${props => props.theme.layout.maxWidth};
   display: flex;
   flex-wrap: wrap;
-`;
-
-const SearchBlockHero = styled.div`
-  width: 100%;
-  background: ${props => props.theme.color.brand.primary};
-`;
-
-const SearchHeroForm = styled.form`
-  position: relative;
-  max-width: 400px;
-  margin: auto;
-  padding-top: 1rem;
-  padding-bottom: 2rem;
-  display: flex;
 `;
 
 const SearchInput = styled(Input)`
@@ -48,18 +37,6 @@ const SearchButton = styled(BaseButton)`
   border: none;
   margin-left: -3px;
   border-radius: 0 3px 3px 0;
-`;
-
-const SearchInputLabel = styled(InputLabel)`
-  position: absolute;
-  height: 1px;
-  width: 1px;
-  overflow: hidden;
-  word-wrap: normal;
-  margin: -1px;
-  padding: 0;
-  clip: rect(0, 0, 0, 0);
-  border: 0;
 `;
 
 function SearchContent() {
